+++ src/platform_impl/android/mod.rs
@@ -12,12 +12,12 @@ use std::{
     time::{Duration, Instant},
 };
 
+use crate::utils::Lazy;
 use android_activity::input::{InputEvent, KeyAction, Keycode, MotionAction};
 use android_activity::{
     AndroidApp, AndroidAppWaker, ConfigurationRef, InputStatus, MainEvent, Rect,
 };
 use log::{debug, trace, warn};
-use once_cell::sync::Lazy;
 
 use crate::{
     cursor::Cursor,