+++ src/platform_impl/macos/mod.rs
@@ -5,6 +5,7 @@ mod app;
 mod app_delegate;
 mod cursor;
 mod event;
+mod event_handler;
 mod event_loop;
 mod ffi;
 mod menu;